# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "nadeemshaik1171998@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/addtocart": {
            "post": {
                "summary": "Add Book to cart",
                "description": "User Can Add Book to cart ",
                "operationId": "163adcc5a0eb5ada20a9b7e57888a99d",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "book_id"
                                ],
                                "properties": {
                                    "book_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book added to Cart Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/deleteBookByCartId": {
            "post": {
                "summary": "Delete the book from cart",
                "description": " Delete cart ",
                "operationId": "ddf2bd9696caf3ef09cffc5794f1a78d",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book deleted Sucessfully from cart"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/getAllBooksByUserId": {
            "get": {
                "summary": "Get All Books Present in Cart",
                "description": " Get All Books Present in Cart ",
                "operationId": "d25cc9ca2acf117c491545cbb31a5ce8",
                "requestBody": {},
                "responses": {
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/increamentBookQuantityInCart": {
            "post": {
                "summary": "Add Quantity to Existing Book in cart",
                "description": " Add Book Quantity  in cart",
                "operationId": "f9c456ebe74f43ddd5838d8baf159fc7",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book Quantity increament Successfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/decreamentBookQuantityInCart": {
            "post": {
                "summary": "Decrese the Quantity of Existing Book in cart",
                "description": " Decrese Book Quantity  in cart",
                "operationId": "ade721dddcaed0dbc298fe6b9f01bb5c",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book Quantity decreament Successfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/addBookToCartBywishlist": {
            "post": {
                "summary": "Add the book to Cart from wishlist",
                "description": " Add to cart ",
                "operationId": "9fc45aa616796d24837d73534bf2e4d3",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "wishlist_id"
                                ],
                                "properties": {
                                    "wishlist_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book added to Cart Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}